requirements.txt
flask
flask_cors
numpy
streamlit
requests
pandas 
numpy
matplotlib
torch
yfinance
keras
newsapi
textblob
sklearn
schedule
tensorflow
transformer